src/gcp-talk-to-your-data-gui/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-secret-manager>=2.16.0
db-dtypes>=1.1.0
openai>=1.0.0
pyarrow>=14.0.0
